requirements.txt
pyserde
annotated-doc==0.0.4
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
croniter==6.0.0
db==0.1.1
distro==1.9.0
fastapi==0.128.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==2.15.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
rq==2.6.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0